+++ src/cli/studioctl-server/Program.cs
@@ -3,6 +3,7 @@ using Altinn.Studio.StudioctlServer.Discovery;
 using Altinn.Studio.StudioctlServer.HostBridge;
 using Altinn.Studio.StudioctlServer.Platform;
 using Altinn.Studio.StudioctlServer.Studioctl;
+using Altinn.Studio.StudioctlServer.Topology;
 
 namespace Altinn.Studio.StudioctlServer;
 
@@ -30,6 +31,7 @@ internal static class Program
 
         builder.Services.AddDiscoveryServices(builder.Configuration);
         builder.Services.AddStudioctlServices();
+        builder.Services.AddTopologyServices(builder.Configuration);
         builder.Services.AddHostBridgeServices(builder.Configuration);
 
         builder.WebHost.ConfigureKestrel((context, options) => IpcListener.Configure(context.Configuration, options));
